refactor: split About/Projects page components per SRP
About.tsx (568 lines) and Projects.tsx (509 lines) mixed page layout,
business logic, and rendering for every section in one file each,
plus project-card.tsx/project-details.tsx duplicated the same
carousel/status-icon JSX. Split each concern into its own component:

- About.tsx -> 9 section components under components/custom/about/
  (profile, personal info, hard skills, experience, education,
  certificates, recommendations, languages, soft skills). About.tsx
  is now a thin orchestrator (~60 lines).
- Projects.tsx -> useProjectFilters (filter state/derivation) and
  useInfiniteReveal (generic IntersectionObserver reveal, reusable
  beyond this list) hooks, plus ProjectFilterBar/
  ProjectTechnologyFilterPopover/ProjectActiveFilterBadges/ProjectList/
  ProjectEmptyState components. Projects.tsx drops to ~55 lines and no
  longer holds filter logic directly (DIP: depends on the hook
  abstraction, not inline state).
- project-details.tsx -> ProjectImageCarousel (was duplicated almost
  identically between the version-tab and plain-image code paths),
  ProjectVersionTabs, and ProjectLinks. Giving each carousel instance
  its own local api/currentImage state (previously one shared state
  wired to every version tab's carousel via the same setApi) also
  fixes cross-tab pagination state bleeding between versions.
- New shared ProjectStatusIcons + ProjectDetailHeader remove the
  duplicated status-icon block between project-card.tsx and
  ProjectDetail.tsx.
- Shared scrollReveal/mountReveal motion helpers replace three
  separate copies of the same reveal-animation object shape.

No behavior changes intended; verified via tsc, next build, and
Playwright checks of /about, /projects (filter + infinite scroll),
and a desktop + mobile project detail page.
