+++ src/views/About.tsx
@@ -1,130 +1,40 @@
 'use client'
 
-import { motion } from 'framer-motion'
-import parse from 'html-react-parser'
-import {
-  Award,
-  BookOpen,
-  Briefcase,
-  Building2,
-  Calendar,
-  Clock,
-  Code,
-  ExternalLink,
-  Globe,
-  GraduationCap,
-  Heart,
-  Lightbulb,
-  Mail,
-  MapPin,
-  MessageSquare,
-  PauseCircle,
-  Phone,
-  PlayCircle,
-  Trophy,
-  User,
-  Zap,
-} from 'lucide-react'
 import { useSearchParams } from 'next/navigation'
-import { useEffect, useState } from 'react'
+import { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
-import CarouselPagination from '@/components/custom/carousel-pagination'
+import { AboutCertificatesSection } from '@/components/custom/about/about-certificates-section'
+import { AboutEducationSection } from '@/components/custom/about/about-education-section'
+import { AboutExperienceSection } from '@/components/custom/about/about-experience-section'
+import { AboutHardSkillsSection } from '@/components/custom/about/about-hard-skills-section'
+import { AboutLanguagesSection } from '@/components/custom/about/about-languages-section'
+import { AboutPersonalInfo } from '@/components/custom/about/about-personal-info'
+import { AboutProfessionalProfile } from '@/components/custom/about/about-professional-profile'
+import { AboutRecommendationsSection } from '@/components/custom/about/about-recommendations-section'
+import { AboutSoftSkillsSection } from '@/components/custom/about/about-soft-skills-section'
 import Container from '@/components/custom/container'
 import Sidebar from '@/components/custom/custom-sidebar'
-import { HardSkillsSection } from '@/components/custom/hard-skills-section'
 import Header from '@/components/custom/header'
-import { ImageViewer } from '@/components/custom/image-viewer'
-import Section from '@/components/custom/section'
-import SectionItem from '@/components/custom/section-item'
-import TechnologiesSection from '@/components/custom/technology-section'
-import { Badge } from '@/components/ui/badge'
-import { Card, CardContent, CardHeader } from '@/components/ui/card'
-import {
-  Carousel,
-  type CarouselApi,
-  CarouselContent,
-  CarouselItem,
-} from '@/components/ui/carousel'
 import { useUser } from '@/hooks/use-user'
-import { cn } from '@/lib/utils/cn'
-import { formationStatus } from '@/types/FormationType'
-
-const EASE_OUT = [0.16, 1, 0.3, 1] as const
 
 export function About() {
-  const { t, i18n } = useTranslation()
+  const { i18n } = useTranslation()
   const { user } = useUser()
   const searchParams = useSearchParams()
-  const [currentCertificate, setCurrentCertificate] = useState<number>(0)
-  const [api, setApi] = useState<CarouselApi>()
-
-  const reveal = (index: number) => ({
-    initial: { opacity: 0, y: 20 },
-    whileInView: { opacity: 1, y: 0 },
-    viewport: { once: true, amount: 0.2 },
-    transition: {
-      duration: 0.5,
-      delay: Math.min(index * 0.06, 0.3),
-      ease: EASE_OUT,
-    },
-  })
 
   useEffect(() => {
     i18n.changeLanguage(i18n.language)
-
-    if (!api) {
-      return
-    }
-    setCurrentCertificate(api.selectedScrollSnap())
-
-    api.on('select', () => {
-      setCurrentCertificate(api.selectedScrollSnap())
-    })
-  }, [i18n.language, api])
+  }, [i18n, i18n.language])
 
   useEffect(() => {
     const search = searchParams.toString()
     if (!search) {
       window.scrollTo(0, 0)
+      return
     }
-    if (search) {
-      const element = document.getElementById(search)
-      element?.scrollIntoView({ behavior: 'smooth' })
-    }
+    const element = document.getElementById(search)
+    element?.scrollIntoView({ behavior: 'smooth' })
   }, [searchParams])
-
-  const getEducationStatusBadge = (formation: any) => {
-    if (formation.graduated) {
-      return (
-        <Badge className="border-green-300 bg-green-100 text-base text-green-800 hover:bg-green-100 dark:bg-green-900 dark:text-green-300">
-          <Trophy className="mr-1 h-3 w-3" />
-          {t('about.graduated')}
-        </Badge>
-      )
-    }
-
-    if (
-      formation.currentStatus?.toLowerCase().includes('trancado') ||
-      formation.currentStatus?.toLowerCase().includes('deferred')
-    ) {
-      return (
-        <Badge
-          className="border-red-300 bg-red-50 text-base text-red-700 dark:bg-red-950 dark:text-red-300"
-          variant="outline"
-        >
-          <PauseCircle className="mr-1 h-3 w-3" />
-          {t('about.deferred')}
-        </Badge>
-      )
-    }
-
-    return (
-      <Badge className="border-blue-300 bg-blue-100 text-base text-blue-800 hover:bg-blue-100 dark:bg-blue-900 dark:text-blue-300">
-        <PlayCircle className="mr-1 h-3 w-3" />
-        {t('about.inProgress')}
-      </Badge>
-    )
-  }
 
   return (
     <div className="text-foreground">
@@ -132,435 +42,20 @@ export function About() {
       <Container>
         <Header />
         <div className="space-y-8">
-          <Section.Root>
-            <Section.Title className="flex items-center gap-3 text-xl font-semibold text-foreground">
-              <User className="h-5 w-5 !text-mainColor" />
-              {t('about.professionalProfile')}
-            </Section.Title>
-            <Section.Content>
-              <Card>
-                <CardContent>
-                  <p className="text-base leading-relaxed text-foreground">
-                    {parse(user.professionalProfile)}
-                  </p>
-                </CardContent>
-              </Card>
-            </Section.Content>
-          </Section.Root>
-
-          <Section.Root id={t('about.personalInfo')}>
-            <Section.Title className="flex items-center gap-3 text-xl font-semibold text-foreground">
-              <User className="h-5 w-5 !text-mainColor" />
-              {t('about.personalInfo')}
-            </Section.Title>
-            <Section.Content className="grid gap-x-8 gap-y-5 divide-y divide-border sm:grid-cols-2 sm:divide-y-0">
-              {[
-                {
-                  icon: MapPin,
-                  label: t('about.location'),
-                  value: user.location,
-                },
-                {
-                  icon: MapPin,
-                  label: t('about.address'),
-                  value: user.address,
-                },
-                { icon: Mail, label: t('about.email'), value: user.email },
-                { icon: Phone, label: t('about.phone'), value: user.phone },
-              ].map(item => (
-                <div className="flex items-center gap-3" key={item.label}>
-                  <item.icon className="h-4 w-4 shrink-0 !text-mainColor" />
-                  <div>
-                    <p className="mb-1 text-sm font-medium tracking-wide text-muted-foreground">
-                      {item.label}
-                    </p>
-                    <p className="text-base text-foreground">{item.value}</p>
-                  </div>
-                </div>
-              ))}
-            </Section.Content>
-          </Section.Root>
-
-          <Section.Root id={t('about.technologiesTitle')}>
-            <Section.Title className="flex items-center gap-3 text-xl font-semibold text-foreground">
-              <Code className="h-5 w-5 !text-mainColor" />
-              {t('about.technologiesTitle')}
-            </Section.Title>
-            <Section.Content>
-              <Card>
-                <CardContent className="pt-6">
-                  <HardSkillsSection skills={user.hardSkills} />
-                </CardContent>
-              </Card>
-            </Section.Content>
-          </Section.Root>
-
-          <Section.Root id={t('about.experiencesTitle')}>
-            <Section.Title className="flex items-center gap-3 text-xl font-semibold text-foreground">
-              <Briefcase className="h-5 w-5 !text-mainColor" />
-              {t('about.experiencesTitle')}
-            </Section.Title>
-
-            <Section.Content className="flex flex-col gap-2">
-              {user.experiences.map((experience, index) => (
-                <motion.div key={index} {...reveal(index)}>
-                  <Card className="relative transition-[transform,box-shadow] duration-200 ease-out hover:-translate-y-1 hover:shadow-lg p-6 gap-2 flex flex-col">
-                    {experience.logo && (
-                      <img
-                        alt={`${experience.enterprise} logo`}
-                        className="absolute top-6 right-6 h-12 w-12 object-contain rounded-md bg-white p-1"
-                        decoding="async"
-                        loading="lazy"
-                        src={experience.logo}
-                      />
-                    )}
-                    <CardHeader className="p-0 gap-4 flex flex-col">
-                      <Section.Title className="flex items-center gap-2">
-                        <Briefcase className="h-4 w-4 !text-mainColor" />
-                        <h3 className="text-lg font-semibold text-foreground">
-                          {experience.role}
-                        </h3>
-                      </Section.Title>
-                    </CardHeader>
-                    <CardContent className="p-0">
-                      <div className="space-y-3">
-                        <div className="flex items-center gap-4 text-base !text-mainColor dark:!text-mainColor">
-                          <div className="flex items-center gap-1">
-                            <Calendar className="h-3 w-3" />
-                            <span>
-                              {t('about.start')}: {experience.startsDate} -{' '}
-                              {t('about.end')}: {experience.endsDate}
-                            </span>
-                          </div>
-                          <div className="flex items-center gap-1">
-                            <MapPin className="h-3 w-3" />
-                            <span>{experience.location}</span>
-                          </div>
-                        </div>
-
-                        <div className="flex items-center gap-2">
-                          <Building2 className="h-4 w-4 !text-mainColor" />
-                          <span className="text-base text-foreground">
-                            {t('about.enterprise')}:{' '}
-                            <span className="font-medium !text-mainColor dark:!text-mainColor">
-                              {experience.enterprise}
-                            </span>
-                          </span>
-                        </div>
-
-                        <div className="rounded-lg border-l-4 border-mainColor p-4">
-                          <ul className="list-disc ml-5 space-y-1">
-                            {experience.features.map((feature, idx) => (
-                              <li
-                                className="text-base leading-relaxed text-foreground"
-                                key={idx}
-                              >
-                                {parse(feature)}
-                              </li>
-                            ))}
-                          </ul>
-                        </div>
-                      </div>
-                    </CardContent>
-                  </Card>
-                </motion.div>
-              ))}
-            </Section.Content>
-          </Section.Root>
-
-          <Section.Root id={t('about.educationTitle')}>
-            <Section.Title className="flex items-center gap-3 text-xl font-semibold text-foreground">
-              <GraduationCap className="h-5 w-5 !text-mainColor" />
-              {t('about.educationTitle')}
-            </Section.Title>
-            <Section.Content className="space-y-4">
-              {user.formations.map((formation, index) => (
-                <motion.div key={index} {...reveal(index)}>
-                  <Card className="transition-[transform,box-shadow] duration-200 ease-out hover:-translate-y-1 hover:shadow-lg p-6 gap-2 flex flex-col">
-                    <CardHeader className="p-0">
-                      <div className="flex w-full items-start justify-between gap-4">
-                        <div className="flex flex-1 items-center gap-3">
-                          <div className="rounded-lg">
-                            <BookOpen className="h-4 w-4 text-mainColor" />
-                          </div>
-                          <h3 className="flex-1 text-lg font-semibold text-foreground">
-                            {formation.title}
-                          </h3>
-                        </div>
-                        {getEducationStatusBadge(formation)}
-                      </div>
-                    </CardHeader>
-                    <CardContent className="p-0">
-                      <div className="space-y-2">
-                        <div className="flex items-center gap-3 rounded-lg">
-                          <Building2 className="h-4 w-4 !text-mainColor" />
-                          <span className="text-base font-medium text-foreground">
-                            {formation.institution}
-                          </span>
-                        </div>
-
-                        <div className="grid grid-cols-1 gap-4 md:grid-cols-2">
-                          <div className="flex items-center gap-3 rounded-lg">
-                            <Calendar className="h-4 w-4 !text-mainColor" />
-                            <div>
-                              <span className="block text-base font-medium tracking-wide">
-                                {t('about.period')}
-                              </span>
-                              <span className="text-base text-foreground">
-                                {formation.startsDate} - {formation.endsDate}
-                              </span>
-                            </div>
-                          </div>
-
-                          {formation.duration && (
-                            <div className="flex items-center gap-3 rounded-lg p-3">
-                              <Clock className="h-4 w-4 !text-mainColor" />
-                              <div>
-                                <span className="block text-base font-medium tracking-wide">
-                                  {t('about.duration')}
-                                </span>
-                                <span className="text-base text-foreground">
-                                  {formation.duration}
-                                </span>
-                              </div>
-                            </div>
-                          )}
-                        </div>
-
-                        <div
-                          className={cn(
-                            'rounded-lg border-l-4 p-4',
-                            formation.currentStatus ===
-                              formationStatus.COMPLETED && 'border-green-500',
-                            formation.currentStatus ===
-                              formationStatus.DEFERRED && 'border-red-500',
-                            formation.currentStatus ===
-                              formationStatus.IN_PROGRESS && 'border-blue-500'
-                          )}
-                        >
-                          {' '}
-                          <div className="flex items-center gap-2">
-                            <Zap className="h-4 w-4 !text-mainColor" />
-                            <span className="text-base font-medium tracking-wide underline">
-                              {t('about.currentStatus')}:
-                            </span>
-                          </div>
-                          <span className="mt-1 block text-base">
-                            {formation.currentStatus ===
-                              formationStatus.COMPLETED && (
-                              <span className="flex items-center gap-2 font-medium text-green-600">
-                                <Trophy className="h-4 w-4" />
-                                {t('about.completedSuccessfully')}
-                              </span>
-                            )}
-                            {formation.currentStatus ===
-                              formationStatus.DEFERRED && (
-                              <span className="flex items-center gap-2 font-medium text-red-600">
-                                <PauseCircle className="h-4 w-4" />
-                                {t('about.deferred')}
-                              </span>
-                            )}
-                            {formation.currentStatus ===
-                              formationStatus.IN_PROGRESS && (
-                              <span className="flex items-center gap-2 font-medium text-blue-600">
-                                <PlayCircle className="h-4 w-4" />
-                                {t('about.inProgress')}
-                              </span>
-                            )}
-                          </span>
-                        </div>
-                      </div>
-                    </CardContent>
-                  </Card>
-                </motion.div>
-              ))}
-            </Section.Content>
-          </Section.Root>
-
-          <Section.Root id={t('about.certificatesTitle')}>
-            <Section.Title className="flex items-center gap-3 text-xl font-semibold text-foreground">
-              <Award className="h-5 w-5 !text-mainColor" />
-              {t('about.certificatesTitle')}
-            </Section.Title>
-            <Section.Content>
-              <Card className="transition-[transform,box-shadow] duration-200 ease-out hover:-translate-y-1 hover:shadow-lg p-6 gap-2 flex flex-col">
-                <CardHeader className="p-0">
-                  <div className="flex items-center gap-3">
-                    <div className="rounded-lg ">
-                      <Trophy className="h-4 w-4 !text-mainColor" />
-                    </div>
-                    <h3 className="text-lg font-semibold text-foreground">
-                      {user.certificates[currentCertificate].title}
-                    </h3>
-                  </div>
-                </CardHeader>
-                <CardContent className="p-0">
-                  <div className="space-y-4">
-                    <div className="rounded-lg ">
-                      {user.certificates[currentCertificate].description.map(
-                        (description, index) => (
-                          <p
-                            className="text-base leading-relaxed text-foreground"
-                            key={index}
-                          >
-                            {description}
-                          </p>
-                        )
-                      )}
-                    </div>
-                    <div className="flex items-center gap-2">
-                      <Code className="h-4 w-4 !text-mainColor" />
-                      <p className="text-base font-semibold text-foreground">
-                        {t('about.usedTechnologies')}:
-                      </p>
-                    </div>
-                    <div className="flex w-full flex-wrap">
-                      <TechnologiesSection
-                        technologies={
-                          user.certificates[currentCertificate].technologies
-                        }
-                      />
-                    </div>
-                    {user.certificates[currentCertificate].image && (
-                      <Carousel setApi={setApi}>
-                        <CarouselContent className="h-full w-[500px]">
-                          {user.certificates.map((certificate, index) => (
-                            <CarouselItem className="h-full w-full" key={index}>
-                              <ImageViewer
-                                alt={certificate.title}
-                                src={certificate.image}
-                              >
-                                <div className="h-full w-full rounded-lg ">
-                                  <img
-                                    alt={certificate.title}
-                                    className="h-full w-full rounded-lg object-contain"
-                                    decoding="async"
-                                    loading="lazy"
-                                    src={certificate.image}
-                                  />
-                                </div>
-                              </ImageViewer>
-                            </CarouselItem>
-                          ))}
-                        </CarouselContent>
-                      </Carousel>
-                    )}
-                    <div className="py-2 text-center text-base !text-mainColor dark:!text-mainColor">
-                      <CarouselPagination
-                        api={api}
-                        currentImage={currentCertificate}
-                        images={user.certificates.map(
-                          certificate => certificate.image
-                        )}
-                        setCurrentImage={setCurrentCertificate}
-                      />
-                    </div>
-                  </div>
-                </CardContent>
-              </Card>
-            </Section.Content>
-          </Section.Root>
-
-          <Section.Root id={t('about.recommendationsTitle')}>
-            <Section.Title className="flex items-center gap-3 text-xl font-semibold text-foreground">
-              <MessageSquare className="h-5 w-5 !text-mainColor" />
-              {t('about.recommendationsTitle')}
-            </Section.Title>
-            <Section.Content className="space-y-4">
-              {user.recomendations.map((recomendation, index) => (
-                <motion.div key={recomendation.date} {...reveal(index)}>
-                  <Card className="transition-[transform,box-shadow] duration-200 ease-out hover:-translate-y-1 hover:shadow-lg p-6 gap-2 flex flex-col">
-                    <CardHeader className="p-0">
-                      <div className="flex items-center gap-4">
-                        <SectionItem.Avatar
-                          imageUrl={recomendation.linkedinImageUrl}
-                        />
-                        <div className="flex-1">
-                          <h3 className="text-lg font-semibold text-foreground">
-                            {recomendation.name}
-                          </h3>
-                          <p className="text-base !text-mainColor dark:!text-mainColor">
-                            {recomendation.position}
-                          </p>
-                        </div>
-                        <div className="flex items-center gap-2">
-                          <ExternalLink className="h-4 w-4 !text-mainColor" />
-                          <SectionItem.LinkedIn
-                            recommendation={recomendation}
-                          />
-                        </div>
-                      </div>
-                    </CardHeader>
-                    <CardContent className="p-0">
-                      <div className="space-y-3">
-                        <div className="flex items-center gap-2">
-                          <Building2 className="h-4 w-4 !text-mainColor" />
-                          <p className="text-base font-medium text-foreground">
-                            {recomendation.company}
-                          </p>
-                        </div>
-                        <div className="rounded-lg border-l-4 p-4 border-mainColor">
-                          <div className="mb-2 flex items-center gap-2">
-                            <Calendar className="h-4 w-4 !text-mainColor" />
-                            <p className="text-base font-medium tracking-wide !text-mainColor dark:!text-mainColor">
-                              {recomendation.date}
-                            </p>
-                          </div>
-                          <div className="flex gap-2">
-                            <MessageSquare className="mt-1 h-4 w-4 flex-shrink-0 !text-mainColor" />
-                            <p className="text-base italic leading-relaxed text-foreground">
-                              "{recomendation.comments}"
-                            </p>
-                          </div>
-                        </div>
-                      </div>
-                    </CardContent>
-                  </Card>
-                </motion.div>
-              ))}
-            </Section.Content>
-          </Section.Root>
-
-          <Section.Root id={t('about.languagesTitle')}>
-            <Section.Title className="flex items-center gap-3 text-xl font-semibold text-foreground">
-              <Globe className="h-5 w-5 !text-mainColor" />
-              {t('about.languagesTitle')}
-            </Section.Title>
-            <Section.Content>
-              {user.languages && (
-                <div className="flex flex-wrap gap-x-6 gap-y-3">
-                  {user.languages.map((lang, idx) => (
-                    <div className="flex items-center gap-3" key={idx}>
-                      <Badge className="border-mainColor bg-mainColor/10 text-base text-foreground">
-                        {lang.language}
-                      </Badge>
-                      <p className="text-base text-foreground">{lang.type}</p>
-                    </div>
-                  ))}
-                </div>
-              )}
-            </Section.Content>
-          </Section.Root>
-
-          <Section.Root id={'Soft Skills'}>
-            <Section.Title className="flex items-center gap-3 text-xl font-semibold text-foreground">
-              <Heart className="h-5 w-5 !text-mainColor" />
-              {'Soft Skills'}
-            </Section.Title>
-            <Section.Content>
-              <div className="flex flex-wrap gap-2">
-                {user.softSkills.map((skill, index) => (
-                  <Badge
-                    className="gap-1.5 border-mainColor bg-mainColor/10 py-1.5 text-sm text-foreground"
-                    key={index}
-                  >
-                    <Lightbulb className="h-3.5 w-3.5 !text-mainColor" />
-                    {skill}
-                  </Badge>
-                ))}
-              </div>
-            </Section.Content>
-          </Section.Root>
+          <AboutProfessionalProfile profile={user.professionalProfile} />
+          <AboutPersonalInfo
+            address={user.address}
+            email={user.email}
+            location={user.location}
+            phone={user.phone}
+          />
+          <AboutHardSkillsSection skills={user.hardSkills} />
+          <AboutExperienceSection experiences={user.experiences} />
+          <AboutEducationSection formations={user.formations} />
+          <AboutCertificatesSection certificates={user.certificates} />
+          <AboutRecommendationsSection recommendations={user.recomendations} />
+          <AboutLanguagesSection languages={user.languages} />
+          <AboutSoftSkillsSection skills={user.softSkills} />
         </div>
       </Container>
     </div>
